flask/user.py/requirements.txt
fastapi==0.95.0
tortoise-orm==0.21.0
uvicorn==0.22.0
passlib[bcrypt]==1.7.4
pydantic==1.10.0
